un monton de vueltas, principalmente UsuarioServiceImpl y EstudianteServiceImpl

+++ src/main/java/edu/unimagdalena/reservasespacios/dtos/requests/estudiante/EstudianteDTOCreate.java
@@ -1,7 +1,7 @@
 package edu.unimagdalena.reservasespacios.dtos.requests.estudiante;
 
 public record EstudianteDTOCreate(
-        int codEstudiantil,
+        Long codEstudiantil,
         String nombre,
         String correo,
         String contrasena
